If the current sessions gain is -8x bet or more and you get one bet correct, the next bet will be half the amount of current negative bets instead of 1x more than the current negative like it is now.
Lets say you bet 10, you're at a negative 110, you win so the current negative would be 100. The application would ask you to bet 50 instead of 110.

And if it requires to split the bet since the negative was even, for example -12x bet, then you win so it's -11x bet. The application will ask you to bet 6x bet instead of 5.5x as it will be rounded upwards for simplicity.